Clean up quirk, replace old v1 quirk
